Launch a browser from the tray when given /welcome
The AHK installer would run both executables. InnoSetup doesn't seem to
have a way to run multiple things in one [Run] entry, short of a batch
script.

This is necessary for displaying the first run setup. The wrapper is
started first because the launcher expects it and there are concurrency
issues where the wrapper is unable to find FProxy otherwise.