Convert Russian translation to codepage 1251
Checking against the Russian.isl default translations, Cryllic script
uses the Windows 1251 codepage. The LangOptions section is copied from it.

I'm not sure of what the intended way to convert this is, but I ended up
pasting the source file into notepad, then back into Notepad++ with the
codepage set.