Use custom hardcoded Firefox path
A user running XP says the tray does not detect their Firefox. Rather than
write an entire feature for this strange use case which involves a
customized Windows build which replaces IE, and so causes IE to fail to
open, this seems more practical.

This also makes Auto mode fall back to a system URL call to allow the
explicit choice between that and the custom Firefox.